+++ backend/src/main/java/com/example/auction/auction/dto/AuctionListResDto.java
@@ -14,7 +14,7 @@ public class AuctionListResDto {
     private Long auctionId;
     private String mainImageUrl;
     private String title;
-    private Integer currentPrice;
+    private Long currentPrice;
     private int bidCount;
     private LocalDateTime endTime;
 
@@ -23,7 +23,7 @@ public class AuctionListResDto {
     private String sellerNickname;
 
     @QueryProjection
-    public AuctionListResDto(Long auctionId, String mainImageUrl, String title, Integer currentPrice, int bidCount, LocalDateTime endTime, AuctionStatus status, String sellerNickname) {
+    public AuctionListResDto(Long auctionId, String mainImageUrl, String title, Long currentPrice, int bidCount, LocalDateTime endTime, AuctionStatus status, String sellerNickname) {
         this.auctionId = auctionId;
         this.mainImageUrl = mainImageUrl;
         this.title = title;
